@@ -1,7 +1,5 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Pobytne.Client.Extensions;
-using System.Net.Http;
-using System.Security.Claims;
 
 namespace Pobytne.Client.Authentication
 {
@@ -9,26 +7,28 @@ namespace Pobytne.Client.Authentication
     {
         protected override Task HandleRequirementAsync(AuthorizationHandlerContext context, PermitionRequirement requirement)
         {
-            var permition = context.User.Claims.FirstOrDefault(c => c.Type == "Permitions");
-            if (permition == null)
-            {
-				Console.WriteLine("Permition is null");
-				context.Fail(new(this, "Access denied"));
-                Console.WriteLine(context.User.Claims.Count());
-                return Task.CompletedTask;
-            }
+            var Resource = context.Resource as (PermitionEnum, AccessEnum,string)?;
+            if(Resource is not null)
+            {//TODO: possilkat v JWT vsechny pristupy... Pak vyhledat pristup v zavislosti na idModulu
+                var permitions = context.User.Claims.FirstOrDefault(c => c.Type == "Permitions");// pole mermici
 
-            string permitionString = permition.Value;
+                if (permitions == null)
+                {
+				    context.Fail(new(this, "Access denied"));
+                    return Task.CompletedTask;
+                }
+
+                string permitionString = permitions.Value;
 
 
-            var Resource = context.Resource as (PermitionEnum, AccessEnum)?;
-            if (Resource is not null && requirement.GetAccess(Resource.Value, permitionString))
-            {
-                context.Succeed(requirement);
-            }
-            else
-            {
-                context.Fail(new(this, "Access denied"));
+                if (Resource is not null && requirement.GetAccess((Resource.Value.Item1,Resource.Value.Item2), permitionString))
+                {
+                    context.Succeed(requirement);
+                }
+                else
+                {
+                    context.Fail(new(this, "Access denied"));
+                }
             }
 
             return Task.CompletedTask;
