@@ -1,4 +1,4 @@
-﻿@page "/login"
+﻿@page "/Login"
 @attribute [AllowAnonymous]
 @inject AuthenticationService authService
 @inject NavigationManager navManager
