@@ -1,7 +1,7 @@
 ﻿@if(moduleCopy is not null && module is not null)
 {
     <EditForm Model="@moduleCopy" OnValidSubmit=OnValidSubmitHandler>
-	    <DataAnnotationsValidator />
+        <DataAnnotationsValidator />
 
         <div class="row align-items-start">
             <div class="col">
